Move plan-approval options into services/plan-approval; apply to omnibus
Plan-approval options are not direct-mode-specific, so relocate
TokenWhitelistingOption, GasPrefundingOption and WalletActivationOption
from services/direct into services/plan-approval/options, and add a
shared buildCustodyPlanApprovalService factory that assembles them
(whitelisting + gas prefunding, plus wallet activation prepended only on
Hedera-style networks, probed once).

Both custody-backed modes now use the factory: direct mode as before,
and omnibus mode — which previously ran a bare PlanApprovalServiceImpl
with no options. Contract mode (on-chain operator, no custody wallets)
is unchanged. The options import the direct-mode abstractions
(CustodyProvider/AccountMapping/AssetStore/WalletActivator) as before;
direct no longer depends on plan-approval.
