+++ src/services/direct/index.ts
@@ -10,7 +10,4 @@ export * from './mapping-validator'
 export * from './mapping-config'
 export * from './wallet-resolver'
 export * from './local-submit-signer'
-export * from './gas-prefunding-option'
-export * from './token-whitelisting-option'
 export * from './wallet-activation'
-export * from './wallet-activation-option'
