+++ src/services/plan-approval/index.ts
@@ -1,3 +1,4 @@
 export * from "./option";
 export * from "./plan-introspection";
 export * from "./configurable-plan-approval-service";
+export * from "./options";
