+++ tests/token-whitelisting.test.ts
@@ -1,4 +1,4 @@
-import { TokenWhitelistingOption } from "../src/services/direct/token-whitelisting-option";
+import { TokenWhitelistingOption } from "../src/services/plan-approval/options/token-whitelisting-option";
 import { tokenStandardRegistry } from "../src/integrations/token-standards/registry";
 import { supportsWhitelisting, WhitelistParty } from "@owneraio/finp2p-ethereum-adapter-contract";
 import { IntrospectedPlan } from "../src/services/plan-approval";
